refactor: switch to new modular scrape architecture
- Renamed old monolithic scrape.rs to scrape_old.rs
- Now using scrape/ module with organized code.rs
- Added execute() function to scrape/mod.rs for backward compatibility
- Removed scrape_new.rs (no longer needed)
- Deprecated query functionality (should use 'patina ask')
- Successfully compiles and ready for use
